prototype1/requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.2a2
gunicorn==19.6.0

# Static and Media Storage
# ------------------------------------------------
boto==2.45.0
django-storages-redux==1.3.2


# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.6.1

# Raven is the Sentry client
# --------------------------
raven==5.32.0


